+++ src/include/constants.h
@@ -42,9 +42,9 @@ typedef signed char sign_t;
 #define M_LOGPI 1.1447298858494002 /**< \f$\log(\pi)\f$ */
 #endif
 
-#define CASIMIR_hbar    1.0545718e-34   /**< reduced Planck constant \f$\hbar\f$ [m² kg / s] */
-#define CASIMIR_hbar_eV 6.582119514e-16 /**< reduced Planck constant \f$\hbar\f$ [eV s/rad] */
-#define CASIMIR_kB      1.38064852e-23  /**< Boltzman constant \f$k_\mathrm{B}\f$ [m² kg / ( K s² )] */
-#define CASIMIR_c       299792458.      /**< speed of light \f$c\f$ in vacuum [m/s] */
+#define CAPS_hbar    1.0545718e-34   /**< reduced Planck constant \f$\hbar\f$ [m² kg / s] */
+#define CAPS_hbar_eV 6.582119514e-16 /**< reduced Planck constant \f$\hbar\f$ [eV s/rad] */
+#define CAPS_kB      1.38064852e-23  /**< Boltzman constant \f$k_\mathrm{B}\f$ [m² kg / ( K s² )] */
+#define CAPS_c       299792458.      /**< speed of light \f$c\f$ in vacuum [m/s] */
 
 #endif