+++ src/tests/test_Lambda.h
@@ -1,6 +1,6 @@
 #ifndef TEST_LAMBDA_H
 #define TEST_LAMBDA_H
 
-int test_casimir_lnLambda(void);
+int test_caps_lnLambda(void);
 
 #endif