+++ src/tests/test_fresnel.h
@@ -1,6 +1,6 @@
 #ifndef TEST_FRESNEL_H
 #define TEST_FRESNEL_H
 
-int test_casimir_fresnel(void);
+int test_caps_fresnel(void);
 
 #endif